got exactly one albume into our db
