+++ server/db/realData.js
@@ -2,7 +2,7 @@ const app = require('../app');
 const axios = require('axios');
 const res = require('express/lib/response');
 
-const sampleData = async (release) => {
+const sampleData = async () => {
   try {
     const response = (
       await axios.get(`https://api.discogs.com/releases/8928025`)
@@ -25,7 +25,7 @@ const sampleData = async (release) => {
       rating: response.community.rating.average,
       availableInventory: response.num_for_sale,
     };
-    console.log(album);
+    // console.log(album);
     return album;
   } catch (error) {
     console.log(error.code);
@@ -37,24 +37,24 @@ const getAlbumDetails = async (artist, title) => {
     const searchResults = await axios.get(
       `https://api.discogs.com/database/search?q=artist=${artist}&title=${title}&key=${process.env.DISCOGS_KEY}&secret=${process.env.DISCOGS_SECRET}`
     );
-    const response = (
-      await axios.get(searchResults.data.results[0].resource_url)
-    ).data;
+    // const response = (
+    //   await axios.get(searchResults.data.results[0].resource_url)
+    // ).data;
     const album = {
-      format: response.formats[0].name,
-      albumName: response.title,
+      // format: response.formats[0].name,
+      // albumName: response.title,
       albumArt: searchResults.data.results[0].cover_image,
-      artistName: response.artists[0].name,
-      genre: response.genres[0],
-      style: response.styles[0],
-      year: response.year,
-      price: response.lowest_price,
-      albumDetails: response.notes,
-      trackList: response.tracklist.map((x) => {
-        return { track: x.position, title: x.title };
-      }),
-      rating: response.community.rating.average,
-      availableInventory: response.num_for_sale,
+      // artistName: response.artists[0].name,
+      // genre: response.genres[0],
+      // style: response.styles[0],
+      // year: response.year,
+      // price: response.lowest_price,
+      // albumDetails: response.notes,
+      // trackList: response.tracklist.map((x) => {
+      //   return { track: x.position, title: x.title };
+      // }),
+      // rating: response.community.rating.average,
+      // availableInventory: response.num_for_sale,
     };
     console.log('--within trycatch--', album);
     return album;
@@ -63,8 +63,8 @@ const getAlbumDetails = async (artist, title) => {
   }
 };
 
-console.log(getAlbumDetails('The Clash', 'London Calling'));
+// const realData = getAlbumDetails('The Clash', 'London Calling');
+// console.log('return value--------', realData);
+// console.log(sampleData());
 
-const realData = getAlbumDetails('The Clash', 'London Calling');
-
-module.exports = realData;
+// module.exports = realData;
